Move ShellStateListener to only test class that uses it
ShellStateListener is only used in a single test class, so move it there
from the to-be-removed UITestCase.
